bot/requirements.txt
tda-api==1.5.2
boto3==1.20.44